requirements.txt
streamlit
opencv-python 
pillow